Updated import-csv.py script and renamed constraints in schema.sql.
The primary change in the script is a logic modification to better categorize authors, coauthors, and illustrators by their roles. This allows for separate dataframes to be built for those that are both author and illustrator, exclusively an author, and exclusively an illustrator. A function has also been added to insert distinct artists into the database. The dataset for game copies has also been updated to account for availability based on the contained 'nature' column. Meanwhile, in the schema.sql file, auto-generated constraint names have been renamed for better readability and understanding.