Enhance handling of 'publisher' field and restructure SQL operations
This update introduces the use of distinct publishers during the CSV data import process. This involves creating a unique dataframe for publishers and implementing SQL operations to handle the insertion of these publishers. Additionally, the 'publisher' field is now attached to the 'games_copy' dataframe and its NULL values are filled with "None". It also includes improved SQL code structure for inserting distinct artists and clean up on some lines of the code.